requirements.txt
pygame-ce>=2.4.0
numpy>=1.24.0
pyyaml>=6.0
